requirements.txt
alembic==1.8.1
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.88.0
flake8==6.0.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
